More fixes for the servant issue (#3).
I didn't catch all of them at first, sorry about that!
Changed the one I missed in English, and figured I can apply this fix
to other languages as well because I'm almost completely certain that
the "R" transliteration doesn't work in any language at all and is just
a mistake on Witch-Hunt's part. I went for "Люнон" in Russian instead of
"Лунон" because it makes the connection to "Люцифер" clearer.